refactor: relocate test case and added new utility test case
