+++ flexmeasures/api/v3_0/tests/test_api_v3_0_users_fresh_db.py
@@ -58,3 +58,32 @@ def test_user_reset_password(
             "reset your password:\n\n%sreset/" % request.host_url
             in pwd_reset_instructions.body
         )
+
+
+@pytest.mark.parametrize(
+    "requesting_user, expected_status_code, user_to_update, expected_role",
+    [
+        ("test_prosumer_user_2@seita.nl", 403, 8, [3, 4]),  # account-admin user
+        # ("test_admin_reader_user@seita.nl", 403, 8, [3]),  # admin reader user
+        # ("test_consultant@seita.nl", 403, 8, [3]),  # consultant user
+        ("test_admin_user@seita.nl", 200, 8, [3, 4]),  # admin user
+    ],
+    indirect=["requesting_user"],
+)
+def test_auth_user_role_modification_permission(
+    client,
+    setup_api_fresh_test_data,
+    requesting_user,
+    expected_status_code,
+    user_to_update,
+    expected_role,
+):
+    patch_user_response = client.patch(
+        url_for("UserAPI:patch", id=user_to_update),
+        json={"flexmeasures_roles": expected_role},
+    )
+
+    print("Server responded with:\n%s" % patch_user_response.data)
+    assert patch_user_response.status_code == expected_status_code
+    if patch_user_response.status_code == 200:
+        assert patch_user_response.json["flexmeasures_roles"] == expected_role
